+++ tests/assets/namespacepackages/locationone/nestednamespace/foo/alpha/blue/one.py
@@ -1,0 +1,6 @@
+import itertools
+import urllib.request
+
+import pytest
+
+from . import two