+++ tests/assets/namespacepackages/locationtwo/nestednamespace/foo/alpha/green/one.py
@@ -1,0 +1,3 @@
+from nestednamespace.foo.alpha.blue import one
+
+from ....bar.beta import orange